Remove useless defer Move test code
