+++ locations/location_mgr.go
@@ -19,7 +19,6 @@ package locations
 import (
 	"encoding/json"
 	"github.com/ystia/yorc/v4/locations/adapter"
-	"github.com/ystia/yorc/v4/prov/hostspool"
 	"path"
 	"time"
 
@@ -75,24 +74,6 @@ func NewManager(cfg config.Configuration) (Manager, error) {
 		}
 		locationMgr = &locationManager{cc: client}
 		locationMgr.hpAdapter = adapter.NewHostsPoolLocationAdapter(client)
-	}
-
-	return locationMgr, nil
-}
-
-// NewManagerWithSSHFactory creates a Location Manager with a given ssh factory
-//
-// Currently this is used for testing purpose to mock the ssh connection.
-func NewManagerWithSSHFactory(cfg config.Configuration, sshClientFactory hostspool.SSHClientFactory) (Manager, error) {
-
-	var locationMgr *locationManager
-	if locationMgr == nil {
-		client, err := cfg.GetConsulClient()
-		if err != nil {
-			return locationMgr, err
-		}
-		locationMgr = &locationManager{cc: client}
-		locationMgr.hpAdapter = adapter.NewHostsPoolLocationAdapterWithSSHFactory(client, sshClientFactory)
 	}
 
 	return locationMgr, nil
